requirements.txt
django == 5.0.6
djangorestframework == 3.15.2
python-dotenv == 1.0.1

mysql == 0.0.3
mysqlclient ==  2.2.4
cryptography == 42.0.8
